pyproject.toml
[project]
name = "hydra-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "pydantic>=2.11.9",
    "torch>=2.8.0",
]
